Refactor reader
