+++ package.json
@@ -36,10 +36,10 @@
     "@digitalcredentials/vc": "^9.0.1",
     "@digitalcredentials/vc-data-model": "^2.0.0",
     "alchemy-sdk": "^3.4.7",
-    "c-kzg": "^4.0.1",
     "caip": "^1.1.1",
     "dotenv": "^16.4.5",
     "ethers": "^6.13.4",
-    "kzg-wasm": "^0.5.0"
+    "kzg-wasm": "^0.5.0",
+    "tsx": "^4.19.2"
   }
 }
