+++ src/index.ts
@@ -1,2 +1,1 @@
-export { isRevoked } from './StatusCheck.ts';
-export { sum } from './sum.ts';
+export { isRevoked } from './lib/StatusCheck.ts';
